Extract MappedPageSource and MappedRecordSet to toolkit
