+++ lib/trino-plugin-toolkit/src/main/java/io/trino/plugin/base/MappedPageSource.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.split;
+package io.trino.plugin.base;
 
 import com.google.common.primitives.Ints;
 import io.trino.spi.Page;
