+++ testing/trino-testing/src/main/java/io/trino/testing/tpch/TpchIndexProvider.java
@@ -14,6 +14,7 @@
 package io.trino.testing.tpch;
 
 import com.google.common.collect.ImmutableList;
+import io.trino.plugin.base.MappedRecordSet;
 import io.trino.plugin.tpch.TpchColumnHandle;
 import io.trino.spi.connector.ColumnHandle;
 import io.trino.spi.connector.ConnectorIndex;
@@ -25,7 +26,6 @@ import io.trino.spi.connector.RecordSet;
 import io.trino.spi.predicate.NullableValue;
 import io.trino.spi.predicate.TupleDomain;
 import io.trino.spi.type.Type;
-import io.trino.split.MappedRecordSet;
 
 import java.util.ArrayList;
 import java.util.List;
